# Optimize zip utility (2x speed increase)