**When merged this pull request will:**
- Optimize zip utility for 2x speed increase using:
  - `BufReader` and `BufWriter`
  - Direct copying without buffer

Test results on 2 of our random mods with `debug` build:

Old   | New  | Zip Size
----- | ---- | --------
28 s  | 14 s | 11.2 MB
122 s | 62 s | 50.6 MB

Deflate implementation in zip-rs is still terribly slow. Bzip2 is very fast, but does not natively work on Windows.