@jonpas 2019-06-14
I just did something way better, now getting 7z-like speeds. Will be in another PR. :)